requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
pypdf
pytesseract
Pillow
sentence-transformers
chromadb
openai
python-dotenv